+++ workspace/TS100/Core/Drivers/TipThermoModel.h
